refactor: WinningNumber를 WinningLotto로 변경
