+++ src/js/models/WinningLottoCounter.js
@@ -1,0 +1,13 @@
+export default class WinningLottoCount {
+  #winningCounts;
+  #winningLotto;
+  constructor() {
+    this.#winningCounts = {
+      '1th': 0,
+      '2th': 0,
+      '3th': 0,
+      '4th': 0,
+      '5th': 0,
+    };
+  }
+}
